AP_Scripting: remove use of uint32_t_ud from serial bindings
Avoids the hassle of having to convert results to regular numbers in
scripts. We trust the user will not be processing 2GiB of data in one
call or attempting to go past 2 gigabaud.

+++ libraries/AP_Scripting/AP_Scripting_SerialAccess.h
@@ -14,13 +14,13 @@ public:
 
     void begin(uint32_t baud);
 
-    size_t write(uint8_t c);
+    int32_t write(uint8_t c);
     size_t write(const uint8_t *buffer, size_t size);
 
     int16_t read(void);
     ssize_t read(uint8_t *buffer, uint16_t count);
 
-    uint32_t available(void);
+    int32_t available(void);
 
     void set_flow_control(enum AP_HAL::UARTDriver::flow_control fcs);
